Restore recent input between restarts
If someone quits the app and restarts in soon afterwards (within ten
minutes) the app will prepopulate the bill field with the previous
amount--though the tip percent will be set to their user default, *not*
the tip percent they last selected.

To be more precise, when the app starts (or restarts), if the last
change to the bill field was less than ten minutes ago, the app will
restore that value. So if you input a value and let the app idle for
longer than ten minutes, then quit and immediately restart the app, the
previous value will not be restored.